Bug 1252737 - use size_t instead of uint32_t for InitEncode(). r=jesup
Although 3rd parameter VideoEncodeer::InitEncode() is defiend as size_t, WebrtcIMXH264VideoCodec and WebrtcMediaCodecVP8VideoCodec uses uint32_t.  This will cause build error for 64bit platform.

MozReview-Commit-ID: 20NgE5e7Cuz

--HG--
extra : rebase_source : 6da7b73504843b1062daff2b6977743114c2591e

+++ media/webrtc/signaling/src/media-conduit/WebrtcOMXH264VideoCodec.h
@@ -40,7 +40,7 @@ public:
 
   virtual int32_t InitEncode(const webrtc::VideoCodec* aCodecSettings,
                              int32_t aNumOfCores,
-                             uint32_t aMaxPayloadSize) override;
+                             size_t aMaxPayloadSize) override;
 
   virtual int32_t Encode(const webrtc::I420VideoFrame& aInputImage,
                          const webrtc::CodecSpecificInfo* aCodecSpecificInfo,
